modified:   app/views/home/index.html.erb

+++ db/migrate/20141230041156_add_code_to_promotion.rb
@@ -1,6 +1,6 @@
 class AddCodeToPromotion < ActiveRecord::Migration
   def change
-    add_column :sdb_imodec_promotions, :code, :integer
+    add_column :sdb_imodec_promotions, :code, :long
 
   end
   def self.down
